post-userリレーション追加

+++ src/main/java/Group/Better/repository/PostRepository.java
@@ -1,6 +1,7 @@
 package Group.Better.repository;
 
 import Group.Better.entity.Post;
+import Group.Better.entity.User;
 import org.apache.ibatis.annotations.*;
 
 import java.util.List;
@@ -9,7 +10,7 @@ import java.util.List;
 public interface PostRepository {
 
     @Insert("insert into posts (title, content, user_id) values (#{title}, #{content}, #{user_id})")
-    void insert(@Param("title") String title, @Param("content") String content, @Param("user_id") int user_id);
+    void insert(@Param("title") String title, @Param("content") String content, @Param("user_id") long user_id);
 
     @Select("select * from posts")
     List<Post> findAll();
